add CatanApp to repository
The app has ability to toggle between code and designer views. Calculate
button is the place where the calculations should go. Right-hand side of
the designer view shows the names of all the buttons/text
entries/outputs. A "label" in matlab is what I'm using to output and an
"edit field" is for input.